+++ apps/commerce-api/src/main/java/com/loopers/domain/point/PointRepository.java
@@ -1,8 +1,10 @@
 package com.loopers.domain.point;
 
+import java.math.BigInteger;
+
 public interface PointRepository {
 
   PointModel get(String userId);
 
-  PointModel charge(String userId, int point);
+  PointModel charge(String userId, BigInteger point);
 }
